Remove race condition on Redis startup for first test.
The flush() code flushes the Redis servers at the start of every test.
This caused a race in the first test: the sleep(args.wait) call comes
AFTER the flush(). Therefore, the first flush() call can come before
the Redis databases have completed initialization. This causes
an error return on the client.flush() call.

The fix simply doesn't do a client.flush() if there are no running
or terminated DB instances. If no DB instances have ever been created,
there is nothing to flush (and the Redis instances might still be
initializing). On the other hand, if DB instances have been created,
we need to flush the Redis instances and we know the Redis instances
have completed initialization.

+++ test/run.py
@@ -273,8 +273,11 @@ def info(msg):
     sys.stdout.flush()
 
 def flush():
-    for client in clients:
-        client.flushall()
+    if len(db_servers) > 0:
+        # Only need to flush if DB servers have actually written to Redis
+        # Otherwise don't call---give Redis servers time to start up
+        for client in clients:
+            client.flushall()
 
 def restartServers(testName):
     stopServers()
